Add CORS headers to ext-auth endpoint
Using the django-cors-headers module, which is recommended by the Django
REST Framework:
<https://www.django-rest-framework.org/topics/ajax-csrf-cors/#cors>. It
only allows cross-origin OPTIONS and POST requests to the ext-auth
endpoint, nothing more.

These headers are required for browsers to allow AJAX requests to a
different page. When doing an AJAX request, the browser first asks via
OPTIONS if it's okay to perform it. If the response is successful, it'll
actually send the request. In both cases, appropriate Access-Control-*
headers need to be set in the response, which the CORS middleware takes
care of.