Implement invites page
Providing a single page that will automatically attempt to open the
session in Drawpile, as well as give options to copy the session link
and download the program. Makes it easier to invite someone without
having to explain how to procure the program and where to punch in the
link.