Point download links at GitHub Releases instead
This is already done for Drawpile 2.2 anyway. Now all the old releases
and artifacts are in GitHub releases too, so we don't need to provide
them ourselves anymore. That will probably mean faster downloads, since
their servers are surely better and are optimized for global content
delivery, and also a nicer interface for browsing through old releases
than the raw nginx file tree.