Support web client and active users in listings
If a session allows joining via web browser, that is now reflected in
the session flags and can be filtered by. Active user counts are listed
with the session if available. Attempting to join the session now leads
to the invite link page where the user is prompted to copy the link,
join through the application or join via web browser, rather than being
direct links. There's also a join button on the right now, rather than
hoping that the user will intuit that clicking on the title lets them
join the session.