Implement authentication Drawpile in the browser
Because that should just use the browser that the user is currently in
to do the authentication, instead of going via the ext-auth API and
requiring them to enter their password all the time. The website now
instead provides a page that can be embedded into an iframe to let the
user log in if necessary and pick a username, after which it will
perform the authentication using the browser cookie.