Adjust confusing buttons on communities page
The buttons to connect and add the community to the browse page are no
longer there, since they don't work unless the user has Drawpile
installed and then only on Windows. Instead there's now a button to go
to the sessions listings and the notice to add the server to Drawpile is
adjusted depending on the operating system.